feat(api): tag and watch (#1834)

+++ ee/scripts/schema/db/init_dbs/postgresql/1.17.0/1.17.0.sql
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS public.projects_conditions
 
 CREATE TABLE IF NOT EXISTS public.tags
 (
-        tag_id               bigint       NOT NULL PRIMARY KEY,
+        tag_id               serial       NOT NULL PRIMARY KEY,
         name                 text         NOT NULL,
         project_id           integer      NOT NULL REFERENCES public.projects (project_id) ON DELETE CASCADE,
         selector             text         NOT NULL,
@@ -58,7 +58,7 @@ CREATE TABLE IF NOT EXISTS events.tags
         session_id bigint  NOT NULL REFERENCES public.sessions (session_id) ON DELETE CASCADE,
         timestamp  bigint  NOT NULL,
         seq_index  integer NOT NULL,
-        tag_id     bigint  NOT NULL REFERENCES public.tags (tag_id) ON DELETE CASCADE,
+        tag_id     integer  NOT NULL REFERENCES public.tags (tag_id) ON DELETE CASCADE,
         PRIMARY KEY (session_id, timestamp, seq_index)
 );
 CREATE INDEX IF NOT EXISTS tags_session_id_idx ON events.tags (session_id);
