+++ ee/scripts/schema/db/init_dbs/postgresql/init_schema.sql
@@ -1156,7 +1156,7 @@ $$
 
             CREATE TABLE public.tags
             (
-                tag_id               bigint       NOT NULL PRIMARY KEY,
+                tag_id               serial       NOT NULL PRIMARY KEY,
                 name                 text         NOT NULL,
                 project_id           integer      NOT NULL REFERENCES public.projects (project_id) ON DELETE CASCADE,
                 selector             text         NOT NULL,
@@ -1171,7 +1171,7 @@ $$
                 session_id bigint  NOT NULL REFERENCES public.sessions (session_id) ON DELETE CASCADE,
                 timestamp  bigint  NOT NULL,
                 seq_index  integer NOT NULL,
-                tag_id     bigint  NOT NULL REFERENCES public.tags (tag_id) ON DELETE CASCADE,
+                tag_id     integer  NOT NULL REFERENCES public.tags (tag_id) ON DELETE CASCADE,
                 PRIMARY KEY (session_id, timestamp, seq_index)
             );
             CREATE INDEX tags_session_id_idx ON events.tags (session_id);
